Linting README.md and all_documents.md
Introduces [pre-commit](https://pre-commit.com) to link Markdown with
[markdownlint-cli2](https://github.com/DavidAnson/markdownlint-cli2) and a few other hooks from
`pre-commit` (`trailing-whitespace`, `end-of-file-fixer` and `check-yaml`).

Line length is set to 120 for `markdownlint-cli2` but ignored for code-blocks and tables. Files with the extension `.md`
and `.markdown` are linted. Issues that can be fixed automatically are.

`.pre-commit-config.yaml` includes a section to enable [`pre-commit.ci`](https://pre-commit.ci) but this requires
linking the organisation to to `pre-commit.ci` and enabling for the repository (I've written a blog post on this the
relevant section is [here](https://ns-rse.github.io/posts/pre-commit-ci/#setup)).

I can work through linting the existing code base in a separate PR but thought it would be useful to get feedback on
the current selection of rules and how happy people are with this before forging ahead.